refactor: styles 폴더 생성 및  styled component 파일 분류
