+++ src/components/LottoPurchaseForm/styles/LottoPurchaseForm.style.js
