+++ src/components/lottoPurchaseResult/LottoIconList.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 
 import LottoIcon from './LottoItem';
 
-import { LottoIconListUl } from './LottoIconList.style';
+import { LottoIconListUl } from './styles/LottoIconList.style';
 class LottoIconList extends Component {
   getLottoIcons() {
     return Array.from({ length: this.props.lottoCount }, (_, idx) => (
