+++ src/components/lottoPurchaseResult/styles/LottoDetailList.style.js
