+++ src/components/lottoPurchaseResult/styles/LottoIconList.style.js
