+++ src/components/lottoPurchaseResult/styles/LottoItem.style.js
