+++ src/components/lottoPurchaseResult/styles/PurchaseResult.style.js
