+++ src/components/lottoPurchaseResult/styles/ToggleButton.style.js
