+++ src/components/lottoRewardResult/styles/RewardResultTable.style.js
