+++ src/components/utils/Flex.js
@@ -1,13 +1,6 @@
 import React, { Component } from 'react';
-import styled from 'styled-components';
 
-const FlexBox = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  margin-top: ${props => props.marginTop};
-`;
+import { FlexBox } from './styles/FlexBox.style';
 
 class Flex extends Component {
   render() {
