+++ src/components/utils/styles/FlexBox.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const FlexBox = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  margin-top: ${props => props.marginTop};
+`;
